use 0-255 int for packed, bool for flipflop
I'm hoping the fingers jerking oddly remotely is caused by floating point rounding

+++ HandOfLesser/src/vrchat/vrchat_osc.h
@@ -45,7 +45,7 @@ namespace HOL::VRChat
 
 		void generateOscTestOutput();
 
-		float encodePacked(float left, float right);
+		int encodePacked(float left, float right);
 
 		float handleInterlacing(float newValue, float oldValue);
